part4 15-23

+++ part4/bloglist/controllers/blog.js
@@ -2,18 +2,24 @@ const blogRoute = require('express').Router()
 const Blog = require('../models/blog')
 
 blogRoute.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogRoute.post('/', async (request, response) => {
   const body = request.body
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
@@ -21,7 +27,7 @@ blogRoute.post('/', async (request, response) => {
 })
 
 blogRoute.get('/:id', async (request, response) => {
-  const blog = await Blog.findById(request.params.id)
+  const blog = await Blog.findById(request.params.id).populate('user', { username: 1, name: 1 })
 
   if (blog) {
     response.json(blog)
@@ -31,6 +37,21 @@ blogRoute.get('/:id', async (request, response) => {
 })
 
 blogRoute.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (!blog) {
+    return response.status(404).json({ error: 'Blog not found' })
+  }
+
+  if (!user) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'Only the creator can delete the blog' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
@@ -47,7 +68,8 @@ blogRoute.put('/:id', async (request, response) => {
 
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
     new: true,
-  })
+  }).populate('user', { username: 1, name: 1 })
+
   response.json(updatedBlog)
 })
